+++ CHANGELOG.md
@@ -7,6 +7,7 @@ HEAD
 * Bump MQTTClient on examples to v0.2.0
 * Removed unnecessary break line in log message
 * Include ESP_LOGE, ESP_LOGD and ESP_LOGV definitions for esp8266 logging
+* Add MQTTClient::getMQTTClientId method
 
 v0.2.0 (2023-05-16)
 ------